xforms-engine: reorganize a couple of parse test suites missed in #213
